requirements.txt
pillow
pydantic